==> notebooks/04_Prep_Business_Data.ipynb <==
# Preparing Business Data for Modeling

Using the data we prepared in previous notebooks, we create a validation set and a feature matrix for use in classification models.

import pandas as pd
import numpy as np

## Load the review features dataset `bus` and the topic matrix `dt_matrix`

bus = pd.read_csv('../data/businesses.csv', compression='gzip')

dt_matrix = np.load('../data/business_doc_topic_matrix.npy')

We will combine the two matrices together into one `features` matrix, so we need to ensure that they are the same shape. 

First we eliminate non-English reviews using the index array we made previously:

correct_index = np.load('../data/bus_eng_index.npy')
bus = bus[bus.index.isin(correct_index)]

bus = bus.reset_index(drop=True)

bus.shape, dt_matrix.shape

We can see that `bus` and `dt_matrix` have the same shape. Let's make a final check for null values before we combine our feature matrices:

bus[bus.isnull().any(axis=1)]

We can drop these reviews as they don't have any feature data other than their text content.

bus.drop([1567253, 1567254], 0, inplace=True)

bus.shape

We also eliminate the topic vectors for those rows and assign the new doc-topic matrix to `right_matrix`:

right_matrix = np.delete(dt_matrix, [1567253, 1567254], 0)

right_matrix.shape

## Create validation set

At this point, we have our two matrices in the same shape. I propose creating a validation set of our data using the following scheme:

 - Reviews with no useful votes will be classified as not useful (0).
 - Reviews with either 1 or 2 useful votes will be held out as a validation set. 
 - Reviews with at least 3 useful votes will be classified as useful (1).
 
This approach intends to control for the unknown factor of page/click count influencing the number of votes a review could get, in addition to the algorithm Yelp uses to filter and display reviews on a specific businesses' page. Following the scheme, we believe that having at least three individuals tag a review as 'Useful' is representative of a consensus, whereas reviews with only 1 or 2 votes may or may not be useful. 

def useful_mapper(x):
    if x == 0:
        return 0
    elif x in (1, 2):
        return "Validation"
    elif x >= 3:
        return 1

bus['usefulness'] = bus['useful'].map(lambda x: useful_mapper(x))

bus['usefulness'].value_counts()

bus.shape[0]

### Use index of validation set to create array of topic vectors for train set

valid_index = bus[bus['usefulness'] == 'Validation'].index

valid_index

right_matrix_valid = right_matrix[valid_index]
right_matrix_valid.shape

right_matrix_train = np.delete(right_matrix, valid_index, axis=0)

right_matrix_train.shape

### Create the validation and train datasets and save to csv

bus_valid_set = bus[bus.index.isin(valid_index)]

bus_valid_set.to_csv('../data/businesses_validation.csv', index=False)

bus_train = bus[~bus.index.isin(valid_index)]

bus_train.to_csv('../data/businesses_train.csv', index=False)

bus_train.values.shape

### Create train feature vector

We drop `text`, `useful`, `cool`, and `state` from the feature vectors.

bus_train.drop(['text', 'useful', 'cool', 'state'], 1, inplace=True)

left_matrix_train = bus_train[bus_train.columns[:-1]].values

left_matrix_train.shape, right_matrix_train.shape

del bus, bus_train, correct_index, dt_matrix, right_matrix

features = np.hstack((left_matrix_train, right_matrix_train))

np.save('../data/businesses_train_features.npy', features)

### Create validation feature vector

bus_valid = pd.read_csv('../data/businesses_validation.csv')

bus_valid.columns

bus_valid.drop(['text', 'useful', 'cool', 'state'], 1, inplace=True)

left_matrix_valid = bus_valid[bus_valid.columns[:-1]].values

left_matrix_valid.shape, right_matrix_valid.shape

del bus, bus_valid, correct_index, dt_matrix, right_matrix

valid_features = np.hstack((left_matrix_valid, right_matrix_valid))

np.save('../data/valid_features.npy', valid_features)

### Create train target vector

bus_train = pd.read_csv('../data/businesses_train.csv')

np.save('../data/business_target.npy', bus_train.usefulness.values)
