==> notebooks/05_Prep_Restaurant_Data.ipynb <==
# Prep data for modeling - Restaurants

## Import modules 

import pandas as pd
import numpy as np

## Load the review metadata set `rests` and the doc-topic matrix `dt_matrix`

rests = pd.read_csv('../data/restaurants.csv', compression='gzip')

dt_matrix = np.load('../data/rests_dt_matrix.npy')

rests.shape, dt_matrix.shape

To correctly combine our metadata and topic vectors into one feature matrix, we need to ensure they are the same shape; we can do this using the index array of the English reviews created in [Notebook 2](02_Text_Preprocessing_and_Tokenizing.ipynb).

correct_index = np.load('../data/rests_eng_index.npy')
rests = rests[rests.index.isin(correct_index)]

rests = rests.reset_index(drop=True)

rests.shape, dt_matrix.shape

## Create validation set following usefulness logic

def useful_mapper(x):
    if x == 0:
        return 0
    elif x in (1, 2):
        return "Validation"
    elif x >= 3:
        return 1

rests['Usefulness'] = rests['useful'].map(useful_mapper)

rests['Usefulness'].value_counts()

## Drop null values

rests[rests.isnull().any(axis=1)]

rests.drop([1798074, 1798075, 2560639], 0, inplace=True)
rests.shape

right_matrix = np.delete(dt_matrix, [1798074, 1798075, 2560639], 0)
right_matrix.shape

## Split out validation and train topic vectors

rests = rests.reset_index(drop=True)

valid_index = rests[rests['Usefulness'] == 'Validation'].index

valid_index

right_matrix_valid = right_matrix[valid_index]
right_matrix_valid.shape

right_matrix_train = np.delete(right_matrix, valid_index, 0)
right_matrix_train.shape

## Create validation and train metadata datasets and save to csv

rests_valid = rests[rests.index.isin(valid_index)]

rests_valid.to_csv('../data/rests_validation.csv', index=False)

rests_train = rests[~rests.index.isin(valid_index)]

rests_train.to_csv('../data/rests_train.csv', index=False)

rests_train.shape, rests_valid.shape

## Create train feature matrix

rests_train.drop(['text', 'useful', 'cool', 'state'], 1, inplace=True)

left_matrix_train = rests_train[rests_train.columns[:-1]].values

left_matrix_train.shape, right_matrix_train.shape

del rests, rests_train, correct_index, dt_matrix, right_matrix

features = np.hstack((left_matrix_train, right_matrix_train))

del right_matrix_train, left_matrix_train

whos

np.save('../data/rests_train_features.npy', features)

## Create validation feature matrix

rests_valid.drop(['text', 'useful', 'cool', 'state'], 1, inplace=True)

left_matrix_valid = rests_valid[rests_valid.columns[:-1]].values

left_matrix_valid.shape, right_matrix_valid.shape

valid_features = np.hstack((left_matrix_valid, right_matrix_valid))

np.save('../data/valid_features_rests.npy', valid_features)

## Create train target vector

rests_train = pd.read_csv('../data/rests_train.csv')

rests_train.Usefulness.value_counts()

np.save('../data/rests_target.npy', rests_train.Usefulness.values)
