==> notebooks/06_Classification_Modeling_Businesses.ipynb <==
# Classification Modeling - Businesses

Fit and cross-validation binary classification models which predict usefulness for non-restaurant business reviews. We choose a logistic regression model that classifies business reviews as useful or not useful with 82% accuracy.
___

## Import modules

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time
import pickle

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier, GradientBoostingClassifier
from sklearn.svm import SVC

from sklearn.metrics import classification_report, confusion_matrix

## Load features and target

features = np.load('../data/businesses_train_features.npy')

target = np.load('../data/business_target.npy')

## Scale, train/test split

X_train, X_test, y_train, y_test = train_test_split(features, target)

ss = StandardScaler()

X_train = ss.fit_transform(X_train)

X_test = ss.transform(X_test)

## Baseline Accuracy

target.mean()

## Logistic Regression

The logistic regression model assumes that the log-odds of the probability of a review being useful review is a linear combination of the features entered into the model. 

lr = GridSearchCV(LogisticRegression(), param_grid={'random_state': [32], 
                                                    'C': [1e-4, 1e-3, 1e-2, 1e-1, 1, 10, 100, 1000], 
                                                    'solver': ['saga'],
                                                    'penalty': ['l2'],
                                                    'n_jobs': [-1],
                                                    'verbose': [1]})

lr.fit(X_train, y_train)

lr.score(X_train, y_train), lr.score(X_test, y_test)

Our model is 82% accurate in predicting whether or not a review had 3+ useful votes or 0 useful votes for our split data. In the [next notebook](https://github.com/gd32/DSI_capstone/blob/master/notebooks/08_Review_Prediction_Businesses.ipynb), I will use the trained model on the validation data to see examples of the model's predictions.

lr.best_params_

Generate confusion matrix:

with open('../models/logisticreg.pkl', 'rb') as m:
    lr = pickle.load(m)

tn, fp, fn, tp = confusion_matrix(y_train, lr.predict(X_train)).ravel()

confusion_matrix

print(classification_report(y_train, lr.predict(X_train), target_names=['Not useful', 'Useful']))

We have a high number of false negatives and a low number of false positives. The classification report shows that we have high specificity but low sensitivity; this is likely because short reviews are generally being classified as not useful even if they speak specifically about a product or service the business provides.
___

Save the model for predictions:

with open('../models/logisticreg.pkl', 'wb') as model:
    pickle.dump(lr, model)

## Random Forest

Random forest classifiers aim to control for the overfitting present in simple decision trees by fitting mulitple trees using a random subset of the feature space.

rf = GridSearchCV(RandomForestClassifier(), param_grid={'random_state': [32],
                                                        'n_estimators': [10, 50, 100],
                                                        'min_samples_split': range(2, 4),
                                                        'min_samples_leaf': range(8, 12),
                                                        'n_jobs': [-1],
                                                        'verbose':[-1]})

%%time
rf.fit(X_train, y_train)

rf.score(X_train, y_train), rf.score(X_test, y_test)

rf.best_params_

### While results of our grid search suggest that model accuracy could be improved by increasing the number of estimators in each tree, I will continue to the prediction stage using the logistic regression model due to the large disparity in fit time (approx. 120 min for a random forest of 100 estimators vs approx. 25 minutes for logistic regression).
